Hotfixes script name on installation
Sinister prefers the URI resource name which, in this case, is a short
name. Passing the --name parameters should work but there is a bug on
jamesqo/sinister#5 that prefers the URI resource name rather than the
parameters given - which means this waits the upstream merge to be
merged.

This commit uses the hotfix for this situation while upstream isn't
merged or accepted.

Related to #2